Restructure user class
