+++ src/client/app/admin/authentication/no-access.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 
-import { User } from '../user.class';
+import { User } from '../user';
 
 import { AuthenticationService } from './authentication.service';
