+++ src/client/app/admin/authentication/unauthorized.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 
-import { User } from '../user.class';
+import { User } from '../user';
 
 @Component({
 	templateUrl: './unauthorized.component.html'
