+++ src/client/app/admin/index.ts
@@ -1,14 +1,13 @@
+export * from './user';
+export * from './authentication';
+export * from './end-user-agreement';
+export * from './user-management';
+export * from './password';
+
 export * from './admin.component';
 
 export * from './admin.service';
 export * from './users.service';
 
-export * from './user.class';
-
 export * from './admin.module';
 export * from './admin-routing.module';
-
-export * from './authentication';
-export * from './end-user-agreement';
-export * from './user-management';
-export * from './password';
