+++ src/client/app/admin/user-management/admin-edit-user.component.ts
@@ -4,8 +4,8 @@ import { Component } from '@angular/core';
 import { ConfigService } from 'app/core';
 import { AlertService } from 'app/shared';
 import { Role } from '../user-management';
+import { User } from '../user';
 
-import { User } from '../user.class';
 import { AdminService } from '../admin.service';
 import { ManageUserComponent } from './manage-user.component';
