+++ src/client/app/admin/user/user.class.ts
@@ -1,4 +1,4 @@
-import { Credentials } from './authentication/credentials.class';
+import { Credentials } from '../authentication/credentials.class';
 
 export class User {
